Split single file up
