+++ lib/fastimage/fastimage_parsing/bmp.rb
@@ -1,0 +1,21 @@
+module FastImageParsing
+  class Bmp # :nodoc:
+    def initialize(stream)
+      @stream = stream
+    end
+  
+    def dimensions
+      d = @stream.read(32)[14..28]
+      header = d.unpack("C")[0]
+
+      result = if header == 12
+                 d[4..8].unpack('SS')
+               else
+                 d[4..-1].unpack('l<l<')
+               end
+
+      # ImageHeight is expressed in pixels. The absolute value is necessary because ImageHeight can be negative
+      [result.first, result.last.abs]
+    end
+  end
+end
