+++ lib/fastimage/fastimage_parsing/heic.rb
@@ -1,0 +1,133 @@
+module FastImageParsing
+  class Heic < ImageBase # :nodoc:
+    def dimensions
+      @max_size = nil
+      @primary_box = nil
+      @ipma_boxes = []
+      @ispe_boxes = []
+      @final_size = nil
+  
+      catch :finish do
+        read_boxes!
+      end
+  
+      @final_size
+    end
+  
+    private
+  
+    def read_boxes!(max_read_bytes = nil)
+      end_pos = max_read_bytes.nil? ? nil : @stream.pos + max_read_bytes
+      index = 0
+  
+      loop do
+        return if end_pos && @stream.pos >= end_pos
+  
+        box_type, box_size = read_box_header!
+  
+        case box_type
+        when "meta"
+          handle_meta_box(box_size)
+        when "pitm"
+          handle_pitm_box(box_size)
+        when "ipma"
+          handle_ipma_box(box_size)
+        when "hdlr"
+          handle_hdlr_box(box_size)
+        when "iprp", "ipco"
+          read_boxes!(box_size)
+        when "ispe"
+          handle_ispe_box(box_size, index)
+        when "mdat"
+          throw :finish
+        else
+          @stream.read(box_size)
+        end
+  
+        index += 1
+      end
+    end
+  
+    def handle_ispe_box(box_size, index)
+      throw :finish if box_size < 12
+  
+      data = @stream.read(box_size)
+      width, height = data[4...12].unpack("N2")
+      @ispe_boxes << { index: index, size: [width, height] }
+    end
+  
+    def handle_hdlr_box(box_size)
+      throw :finish if box_size < 12
+  
+      data = @stream.read(box_size)
+      throw :finish if data[8...12] != "pict"
+    end
+  
+    def handle_ipma_box(box_size)
+      @stream.read(3)
+      flags3 = read_uint8!
+      entries_count = read_uint32!
+  
+      entries_count.times do
+        id = read_uint16!
+        essen_count = read_uint8!
+  
+        essen_count.times do
+          property_index = read_uint8! & 0x7F
+  
+          if flags3 & 1 == 1
+            property_index = (property_index << 7) + read_uint8!
+          end
+  
+          @ipma_boxes << { id: id, property_index: property_index - 1 }
+        end
+      end
+    end
+  
+    def handle_pitm_box(box_size)
+      data = @stream.read(box_size)
+      @primary_box = data[4...6].unpack("S>")[0]
+    end
+  
+    def handle_meta_box(box_size)
+      throw :finish if box_size < 4
+  
+      @stream.read(4)
+      read_boxes!(box_size - 4)
+  
+      throw :finish if !@primary_box
+  
+      primary_indices = @ipma_boxes
+                        .select { |box| box[:id] == @primary_box }
+                        .map { |box| box[:property_index] }
+  
+      ispe_box = @ispe_boxes.find do |box|
+        primary_indices.include?(box[:index])
+      end
+  
+      if ispe_box
+        @final_size = ispe_box[:size]
+      end
+  
+      throw :finish
+    end
+  
+    def read_box_header!
+      size = read_uint32!
+      type = @stream.read(4)
+      [type, size - 8]
+    end
+  
+    def read_uint8!
+      @stream.read(1).unpack("C")[0]
+    end
+  
+    def read_uint16!
+      @stream.read(2).unpack("S>")[0]
+    end
+  
+    def read_uint32!
+      @stream.read(4).unpack("N")[0]
+    end
+  end
+end
