+++ lib/fastimage/fastimage_parsing/psd.rb
@@ -1,0 +1,7 @@
+module FastImageParsing
+  class Psd < ImageBase # :nodoc:
+    def dimensions
+      @stream.read(26).unpack("x14NN").reverse
+    end
+  end
+end
